requirements.txt
fastapi
uvicorn
psycopg2-binary
passlib[bcrypt]
python-dotenv
pydantic
